Modularize routes.ts wave 6: conversations CRUD + SSE send (-550 lines)
Final modularization wave. Extracts the remaining conversation
endpoints from routes.ts:

- routes-modules/conversations-crud.ts (259 lines) — list / create /
  get / patch / delete / batch-delete / messages GET / files GET /
  upload POST, plus the requireConversation + ensureSessionUserInDatabase
  helpers.

- routes-modules/conversations-send.ts (307 lines) — the big
  POST /api/conversations/:id/messages SSE handler. Carries the tier
  check, web-lookup short-circuit (ManagerAgent → IntelligenceAgent),
  memory + admin context injection, and the Ollama streaming pipeline.
  Split from CRUD because it pulls in the entire chat stack.

routes.ts is now 120 lines (was 2,101 at wave 0, 670 at wave 5). All
route registrations are pure register*Routes(app) calls. The 250-350
line per-file rule is satisfied across the routes layer.
